more grandfathering

+++ config_scheduler.go
@@ -10,20 +10,20 @@ import (
 
 // SchedulerConfig holds relevant settings for the scheduler process.
 type SchedulerConfig struct {
-	TaskFinder                    string `bson:"task_finder" json:"task_finder" yaml:"task_finder"`
-	HostAllocator                 string `bson:"host_allocator" json:"host_allocator" yaml:"host_allocator"`
-	DefaultFutureHostPercent      int    `bson:"free_host_fraction" json:"free_host_fraction" yaml:"free_host_fraction"`
-	CacheDurationSeconds          int    `bson:"cache_duration_seconds" json:"cache_duration_seconds" yaml:"cache_duration_seconds"`
-	Planner                       string `bson:"planner" json:"planner" mapstructure:"planner"`
-	TargetTimeSeconds             int    `bson:"target_time_seconds" json:"target_time_seconds" mapstructure:"target_time_seconds"`
-	AcceptableHostIdleTimeSeconds int    `bson:"acceptable_host_idle_time_seconds" json:"acceptable_host_idle_time_seconds" mapstructure:"acceptable_host_idle_time_seconds"`
-	GroupVersions                 bool   `bson:"group_versions" json:"group_versions" mapstructure:"group_versions"`
-	PatchFactor                   int64  `bson:"patch_zipper_factor" json:"patch_factor" mapstructure:"patch_zipper"`
-	PatchTimeInQueueFactor        int64  `bson:"patch_time_in_queue_factor" json:"patch_time_in_queue_factor" mapstructure:"patch_time_in_queue_factor"`
-	CommitQueueFactor             int64  `bson:"commit_queue_factor" json:"commit_queue_factor" mapstructure:"commit_queue_factor"`
-	MainlineTimeInQueueFactor     int64  `bson:"mainline_time_in_queue_factor" json:"mainline_time_in_queue_factor" mapstructure:"mainline_time_in_queue_factor"`
-	ExpectedRuntimeFactor         int64  `bson:"expected_runtime_factor" json:"expected_runtime_factor" mapstructure:"expected_runtime_factor"`
-	GenerateTaskFactor            int64  `bson:"generate_task_factor" json:"generate_task_factor" mapstructure:"generate_task_factor"`
+	TaskFinder                    string  `bson:"task_finder" json:"task_finder" yaml:"task_finder"`
+	HostAllocator                 string  `bson:"host_allocator" json:"host_allocator" yaml:"host_allocator"`
+	DefaultFutureHostPercent      float64 `bson:"free_host_fraction" json:"free_host_fraction" yaml:"free_host_fraction"`
+	CacheDurationSeconds          int     `bson:"cache_duration_seconds" json:"cache_duration_seconds" yaml:"cache_duration_seconds"`
+	Planner                       string  `bson:"planner" json:"planner" mapstructure:"planner"`
+	TargetTimeSeconds             int     `bson:"target_time_seconds" json:"target_time_seconds" mapstructure:"target_time_seconds"`
+	AcceptableHostIdleTimeSeconds int     `bson:"acceptable_host_idle_time_seconds" json:"acceptable_host_idle_time_seconds" mapstructure:"acceptable_host_idle_time_seconds"`
+	GroupVersions                 bool    `bson:"group_versions" json:"group_versions" mapstructure:"group_versions"`
+	PatchFactor                   int64   `bson:"patch_zipper_factor" json:"patch_factor" mapstructure:"patch_zipper"`
+	PatchTimeInQueueFactor        int64   `bson:"patch_time_in_queue_factor" json:"patch_time_in_queue_factor" mapstructure:"patch_time_in_queue_factor"`
+	CommitQueueFactor             int64   `bson:"commit_queue_factor" json:"commit_queue_factor" mapstructure:"commit_queue_factor"`
+	MainlineTimeInQueueFactor     int64   `bson:"mainline_time_in_queue_factor" json:"mainline_time_in_queue_factor" mapstructure:"mainline_time_in_queue_factor"`
+	ExpectedRuntimeFactor         int64   `bson:"expected_runtime_factor" json:"expected_runtime_factor" mapstructure:"expected_runtime_factor"`
+	GenerateTaskFactor            int64   `bson:"generate_task_factor" json:"generate_task_factor" mapstructure:"generate_task_factor"`
 }
 
 func (c *SchedulerConfig) SectionId() string { return "scheduler" }
